Update prime_number_test.py
Fixed an error in the is_prime() function. I originally had the function return True if the input is not a prime and False if the input is a prime, however the function should output the opposite. This error was easy to correct as I only needed to switch the 'True' and 'False' on the return statements.