Update 10001st_prime_7.py
Modified program to find 6th prime. The answer I got from the last version of my code was incorrect. I want to check if my code outputs the correct prime number. In order to do this I want to see if my code outputs the correct value for the 6th prime.