Update special_pythagorean_triplet_9.py
I didn't realise that the 'sqrt()' function was a part of the 'math' library (I thought that it was a part of the Python Standard Library). As a result, I didn't import the math library, which caused an error since the sqrt() function was undefined. This problem was easily fixed by importing the math function and replacing the 'sqrt()' with 'math.sqrt()'