Update special_pythagorean_triplet_9.py
I realised that the reason my problem was
not working properly was because I made a mistake in my 'pythagorean_triplet()' function. I realised that I didn't accout for the fact that not every integer is part of a pythagorean triplet. As a result for an integer which is not part of a pythagorean triplet, e.g. 1, my code would be stuck in a constant loop. Since for the Pythagorean triplet a,b and c, c cannot be greater than a+b, I added an if statement, which, increases the initial integer if c > a+b. This should fix the problem, ensuring that my function doesn't get stuck in an endless loop.