Update special_pythagorean_triplet_9.py
Updated the 'pythagorean_triplet()' function with code that should fix the error that I was having. I have now added an if statement that exits the funciton if no integer triplets exist for that integer. I have also updated some variable names with names that I believe help keep the code clean.