Update special_pythagorean_triplet_9.py
Following my previous update I still had an error in my 'pythagoran_triplet()' function. I quickly realised that this error was caused because I was not updating the variable within the while loop, resulting in an infinite loop. This was easy to fix as I just had to move one line.