Update special_pythagorean_triplet_9.py
realised that there was another error with my code. After using the debugging tool in VS Code, I realised that the error was caused because I was calculating the hypotenuse before increasing the value of 'integer_b'. This resulted in the output value of integer_b being 1 too high. This was very easy to fix.