Update summation_of_primes_10.py
My code was taking a very long time to run and I realised that I may have misunderstood the problem. I originally thought that I had to find the sum of all the primes which have values below 2 000 000, however I now believe that it is the sum that is meant to be below 2 000 000. I have modified my code for this new understanding of the problem.