Update summation_of_primes_10.py
It turns out that my original understanding of the problem was correct. After doing some research I discovered that my implementation of a prime number test is very slow and that this Project Euler problem is designed to help you realise that an implementation similar to my implementation is very inefficient. This problem is designed to make you research more efficient implementations. As a result, I am going to research how to implement a prime number test efficiently and then redo this problem.