Create function_timer.py
I created a new function to help test the relative efficiencies of different versions of the same function. I have created this function because I realised that my prime_number_test function is very inefficient and I wouldn't be surprised if I came across a similar problem in the future. By creating a function to compare different versions of the same function (with respect to their efficiency) I can easily choose the most efficient version of any function I write.