Update prime_number_test_v1.py
I updated the function that I used to check if the input is an integer. When I was updating this funciton I realised that my function didn't stop if the input was not an integer so I had to edit this as well. This was easily fixed by adding the term 'return' on the 6th line.