Update prime_number_test_v3.py
I realised that, since I am now testing if the integer is even, I do not need to start iterating from 2. This is because only even numbers have a factor of 2 and since the function is only iterating through the divisors of odd numbers, there is no need to test for a divisor of 2. This change may not make much of a difference with respect to the efficiency of the function, however I thought that it was a sensible change to make.