Update sieve_of_eratosthenes.py
I realsied that there was a logic error in my implementation of the sieve of eratosthenes. I had not set the program to check if the value of the element corresponding to the highest_known_prime was 'True' but rather if the value of highest_known_prime (which is an integer and therefore never 'True') was 'True'. As a result, the program would never change any of the values of the elements, returning all of the integers within the range as primes. This error was easy to fix as all I had to do was to check the index of the list 'primes'.