Update sieve_of_eratosthenes.py
I realised that the program was not marking the multiples of a prime as 'False' since it was iterating from two each time rather than a multiple of the prime. This error was also easy to fix becasue all I had to do was change the starting value in the range() function.