Update sieve_of_eratosthenes.py
I realised that there was still an error with my code, however it was hard for me to find this error becuase my code was messy. As a result, I made some changes to my code (e.g. creating varaibles to hold specific, calculated values) to make it more clean.