Update sieve_of_eratosthenes.py
I fixed the errors with my code. Firstly, I realised that I had made an error when marking the multiples of a prime. This error was caused because I was iterating to an end value that was too large to be used as an index for the list primes. This prime was easy to fix becuase I realised that since indexing starts at 0, I needed to subract 1 from 'limit_value' rather than add 1. The second error occured where I was creating the list of primes below the limit value. I realised that I misunderstood how you can iterate through a list, which meant that this code was not editing the function at all. I also realised that when I 'popped' an element out of the list, it shortened the list, making the inices to be popped incorrect. To fix this problem, I created a new list containing the indicies of the elements to be popped. I then popped these elements within a for loop, however I reduced the index by the amount of elements popped to account for the changes made to the list.